Fix circuar dependency by moving more types into worker-interface
